Add virtualClusterConfigurationModel
